+++ src/main/java/com/ecole221/banque_api/repository/CompteRepository.java
@@ -9,9 +9,9 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface CompteRepository extends JpaRepository<Compte, Long> {
+public interface CompteRepository extends JpaRepository<Compte, Integer> {
     Optional<Compte> findByNumero(String numero);
     List<Compte> findByClient(Client client);
-    List<Compte> findByClientId(Long clientId);
+    List<Compte> findByClientId(Integer clientId);
     boolean existsByNumero(String numero);
 }
