+++ src/main/java/com/ecole221/banque_api/repository/TypeRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TypeRepository extends JpaRepository<Type, Long> {
+public interface TypeRepository extends JpaRepository<Type, Integer> {
     Optional<Type> findByLibelleIgnoreCase(String libelle);
 }
